% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constMatiGFT.R
\name{constMatiGFT}
\alias{constMatiGFT}
\title{Calculate the implied GFT for a series of option panels, using a square-root frequency scaling and for fixed maturities.}
\usage{
constMatiGFT(panels.list, u.seed, fixed.maturities = c(1/12, 6/12), trade.int,
  output.name, doParallel = 2, ...)
}
\arguments{
\item{panels.list}{A list version of the option panels, as outputted by \code{\link{panelsToList}}}

\item{u.seed}{A vector of frequencies that will be scaled according to a square-root time-to-maturity rule}

\item{fixed.maturities}{A vector of maturities which represent the option trading strategies}

\item{trade.int}{A scalar representing the assumed trading time length. This together with the fixed.maturities defines the interpolating maturities.}

\item{output.name}{The file name where the interpolated igft-s will be saved}

\item{doParallel}{If integer, specifies the number of cores to be used by starting a cluster with the \code{parallel} package. If \code{foreach}, an extant parallel backend will be used with the \code{foreach} package.}

\item{...}{further arguments to be passed to \code{\link{iGFTinterpolate}} and further on to \code{\link{tpsImpliedGFT}}, \code{\link{panelSmoother}} and further to \code{\link{rgl::persp3d}}}
}
\description{
This function takes as input a vector of strings corresponding to option panels and then calculates the implied GFT for a given frequency rule. Then the implied GFT is interpolated for fixed maturities.
}
