% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgedRetRCov.R
\name{hedgedRetRCov}
\alias{hedgedRetRCov}
\title{Realized variance - covariance matrix of hedged option returns.}
\usage{
hedgedRetRCov(igft.list, igft.days, hf.dframe, weekly.grid,
  trade.end.grid = NULL, u.seed, fixed.maturities = c(1/12, 6/12),
  psd.ensure = TRUE, trade.int = 5/252)
}
\arguments{
\item{igft.list}{A list of buy and sell values of the igft (the difference comes from the maturity decreasing over the trading interval + vol. changes) as output by \code{\link{constMatiGFT}}}

\item{igft.days}{A vector containing the days corresponding to igft.list}

\item{hf.dframe}{A data frame containing high frequency futures prices. It has 2+#TTM columns, \code{day}, \code{time} and \code{F_ttm(1)...F_ttm(T)}, \code{T = length(fixed.maturities)}.}

\item{weekly.grid}{The approximately weekly grid of days where we empirically observe option prices (or their transformed values)}

\item{trade.end.grid}{(optional) A grid of days on which trades starting at weekly.grid are supposed to end. Use if weekly.grid has some 'gaps'.}

\item{u.seed}{Vector of base frequencies at which the portfolio values have been calculated. The vector's length is \code{N}}

\item{fixed.maturities}{The maturities of the option portfolios we are mimicking}

\item{psd.ensure}{If set to TRUE (default), the estimated variance covariance matrix is projected onto the space of positive semi-definite matrices as in [paper].}

\item{trade.int}{The trading interval (annualized in business year) over which we hold the option position.}
}
\value{
Array of size \code{(NxT)x(NxT)xS}, where \code{N} is the number of base frequencies, \code{T} is the number of maturities, \code{S} is the length of \code{weekly.grid}.
}
\description{
The function calculates the approximate matrix of realized variance-covariance between hedged GLT portfolios. The hedging frequency is determined by the frequency of stock futures data. If there is missing data in the time series of GLT values, these days are ommitted from hedging. The approximate nature of the result comes from the fact that we use the stochastic integral representation of the hedged GLT portfolio return and use Ito's isometry to build the approximation.
}
\note{
TODO: handle missing data (NA observations in the igft); add paper citation in @param psd.ensure
}
