% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgedReturns.R
\name{hedgedReturns}
\alias{hedgedReturns}
\title{Hedged option returns}
\usage{
hedgedReturns(igft.list, igft.days, hf.dframe, weekly.grid,
  trade.end.grid = NULL, u.seed, fixed.maturities = c(1/12, 6/12),
  trade.int = 5/252)
}
\arguments{
\item{igft.list}{A list of buy and sell values of the igft (the difference comes from the maturity decreasing over the trading interval + vol. changes) as output by \code{\link{constMatiGFT}}}

\item{igft.days}{A vector containing the days corresponding to igft.list}

\item{hf.dframe}{A data frame containing high frequency futures returns. Has 3 columns, day, time and F (which is the futures price)}

\item{weekly.grid}{The approximately weekly grid of days where we empirically observe option prices (or their transformed values)}

\item{trade.end.grid}{(optional) A grid of days on which trades starting at weekly.grid are supposed to end. Use if weekly.grid has some 'gaps'.}

\item{fixed.maturities}{The maturities of the option portfolios we are mimicking}

\item{trade.int}{The trading interval (annualized in business year) over which we hold the option position.}
}
\value{
Returns a data frame containing the starting day of each option position (assumed to be bought at the end of the day) and the corresponding weekly hedged return.
}
\description{
The function calculates returns to trading hedged GLT positions, given a grid of GLT price data and a time series of high frequency stock futures data. The hedging frequency is determined by the frequency of stock futures data. If there is missing data in the time series of GLT values, these days are ommitted from hedging.
}
\details{
This function calculates the hedged option trading returns, using high frequency option data and approximately daily calculated igft.
}
\note{
TODO: Adapt the function so that old calls work as intended, but new calls use daily data for calculating hedged returns; instead of dropping missing-data days from hedging, do wider interpolation; Ensure that with multiple maturities, multiple time series of futures prices are used.
}
