% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGFTinterpolate.R
\name{iGFTinterpolate}
\alias{iGFTinterpolate}
\title{Calculate interpolated constant maturity implied GFTs}
\usage{
iGFTinterpolate(option.panels, mkt.frame, u.seed, mat.vec, sell.offset = 0,
  time.IV = TRUE, k.shrink = 1, ...)
}
\arguments{
\item{option.panels}{A list of Kx2 strike-price matrices}

\item{mkt.frame}{A data-frame showing the available maturities (corresponding to the day where the panels are observed) and corresponding r and q}

\item{u.seed}{A seed which is used to scale the frequencies by \code{\link{uSquareRootMat}}}

\item{mat.vec}{A vector of maturities where the interpolated gft-s are required.}

\item{time.IV}{A logical. If TRUE, then implied volatilities are time extrapolated from the thin-plate-spline. Otherwise transforms are calculated for maturities available in the option panels and the transforms are directly extrapolated.}

\item{...}{Additional arguments passed to \code{\link{impliedGFT}}}

\item{offset}{A scalar/vector that specifies if we want to deviate from the square-root frequency calculation rule}
}
\value{
Returns an UxT matrix containing the interpolated gft at the scaled frequencies. If in u.seed we request to calculate portfolios for a u such that the power does not exist for too many portfolios (or some other error occurs in the \code{\link{spline}()} call), the value \code{-998} will be returned.
}
\description{
This function takes a list of priceMats and a markets, calculates the impleid GFT-s and then interpolatese/extrapolates them to the required maturities.
}
