% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impliedDivergencesWrapper.R
\name{impliedDivergencesWrapper}
\alias{impliedDivergencesWrapper}
\title{Implied divergences from option DB file}
\usage{
impliedDivergencesWrapper(options, pow.vec, t.vec, types, filtering.expression,
  cl = cl, U = NULL, L = NULL, gam.bs = "ds", gam.m = c(1, 0.5),
  bootstrap = F, Nrepl = 1000)
}
\arguments{
\item{options}{data.frame with fields date, k (log(relative strike with respect to underlying)), relMid (relative mid-price with respect to underlying), r (interest rate), q (implied dividend yield), k_stand (k divided by atmIV and sqrt(expiration))}

\item{pow.vec}{vector of divergence powers}

\item{t.vec}{vector of maturities}

\item{types}{vector of characters, length up to 3, can contain "div", "skew" and "quart"}

\item{filtering.expression}{quoted expression to be passed to \code{\link[dplyr]{filter_}}, to select all options take something that evaluates to TRUE, like \code{quote(TRUE)}, to select options by strike, for example \code{quote(k_stand <= 3 & k_stand >= -6)}}

\item{cl}{cluster for parallel calculations}

\item{U}{upper integration limit for pricing in terms of log-strike. U > 0}

\item{L}{lower integration limit for pricing in terms of log-strike. L < 0}

\item{gam.bs}{bs argument to \code{\link[mgcv]{s}}}

\item{gam.m}{m argument to \code{\link[mgcv]{s}}}

\item{bootstrap}{boolean argument to \code{\link{tpsImpliedDivergence}}}

\item{Nrepl}{integer argument to \code{\link{tpsImpliedDivergence}}}
}
\description{
This function takes a data.frame of option data and at each date builds the spline IV surface, based on that estimates prices of divergence, skewness and quarticity swaps
}
