% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linModWrapper.R
\name{constr.lm.estim}
\alias{constr.lm.estim}
\alias{predict.impliedCF.lm}
\title{Estimation of constrained linear model, prediction.}
\usage{
constr.lm.estim(terms, data, R, r, weights = rep(1, nrow(data)), meq = 0)

\method{predict}{impliedCF.lm}(model, newdata)
}
\arguments{
\item{terms}{Formulae for left and right hand side terms (individual components of formula). The explained variables is assumed to be in \code{terms[1]}}

\item{data}{Data frame with right hand side variables}

\item{R}{Matrix of linear restrictions of the form Rb >= r}

\item{r}{Right hand side of linear restriction Rb >= r}

\item{weights}{For some weighted least squares}

\item{meq}{check \code{\link{quadprog::solve.QP}}}

\item{model}{model of class \code{impliedCF.lm}}

\item{newdata}{conforming data frame}
}
\description{
This function takes a formula, a vector of term formulae and estimates a constrained linear model, giving it class \code{impliedCF.lm} so that the predict routine can be used as well.
}
\details{
Wrapper to \code{\link{solve.QP}} for estimating constrained linear models.
}
