% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noArbitragePanel.R
\name{noArbitrageOptionCheck}
\alias{noArbitrageOptionCheck}
\title{Check no-arbitrate conditions in an option panel}
\usage{
noArbitrageOptionCheck(pn, r, q, s0, call = TRUE)
}
\arguments{
\item{params.Q}{A list containing the Q parameter structures}

\item{stateMat}{An Sx2 matrix (with named columns v1,v2) that contains the 
states at which the VIX is to be calculated}

\item{dT}{A positive scalar, which is the interval over which the VIX should 
be calculated}

\item{panel}{data.frame of option prices with fields r (interest rate), logF (log-forward rate), dT (maturity), k (log-strike), relMid (relative price)}
}
\value{
is.no.arbitrage logical, 1 if there is no arbitrage
}
\description{
Given an option panel, the value of the underlying, the interest
  rate for the appropriate maturity and the dividend yield, check 
  no-arbitrage conditions for static payoffs, i.e. that there can be no
  non-positive value portfolio with a strictly positive payoff. This follows
  \code{http://www.fabiomercurio.it/NoArbitrage.pdf}
}
