% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noArbPrices.R
\name{noArbPrices}
\alias{noArbPrices}
\title{No-arbitrage panels}
\usage{
noArbPrices(option.panel)
}
\arguments{
\item{option.panel}{A data-frame with columns relMid = normalized mid-prices (current stock-price = 1), r = annualized interest, logF = log forward price, dT = option maturity}
}
\value{
A no-arbitrage enforced panel with the same structure as the original panel
}
\description{
Calculate minimal distance no-arbitrage prices
}
