% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpsSmoother.R
\name{panelSmoother}
\alias{panelSmoother}
\title{TPS-based smoother for option panels}
\usage{
panelSmoother(option.panels, mkt.frame, gam.bs = "ds", gam.m = c(1, 0.5),
  ...)
}
\arguments{
\item{option.panels}{A list, each element containing a dataframe with normalized mid OTM option prices and LOG strikes. Each panel should correspond to the SAME day.}

\item{...}{further arguments to the smooth construct in \code{\link[mgcf]{gam}}}

\item{mkt.list}{A data-frame, where each row corresponds to the market features associated with the n-th option panel (e.g. maturity, interest rate, etc...)}
}
\value{
A \code{\link{mgcv::gam}} object, interpolating model.
}
\description{
Given a list of option panels, estimates a generalized additive model that can be used for inter- and extrapolation
}
