% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelsToList.R
\name{panelsToList}
\alias{panelsToList}
\title{Convert joined panel into seperate lists, that can be fed to the igft interpolator}
\usage{
panelsToList(panel.names, noisy = TRUE)
}
\arguments{
\item{panel.names}{Vector of panel names that will be loaded and parsed}
}
\value{
Returns a list with components p.list, m.list and days.all. These correspond daily priceMat panel lists, daily market frames and the vector of available days.
}
\description{
Convert joined panel into seperate lists, that can be fed to the igft interpolator
}
