% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateAndSavePanels.R
\name{saveInterpolatedPanels}
\alias{saveInterpolatedPanels}
\title{Calculate the implied GFT for a series of option panels, using a square-root frequency scaling and for fixed maturities.}
\usage{
saveInterpolatedPanels(panels.list, out.mat, output.name, doParallel = 2,
  lib.dir = "test.lib.dir", ...)
}
\arguments{
\item{panels.list}{A list version of the option panels, as outputted by \code{\link{panelsToList}}}

\item{out.mat}{Sx2 matrix, where first column specifies target maturities and the second specifies the number of options available at each maturity (for now they have to be equal).}

\item{output.name}{The file name where the interpolated igft-s will be saved}

\item{doParallel}{If integer, specifies the number of cores to be used by starting a cluster with the \code{parallel} package. If \code{foreach}, an extant parallel backend will be used with the \code{foreach} package.}

\item{k.shrink}{what?!}
}
\description{
This function takes as input a vector of strings corresponding to option panels and then calculates the implied GFT for a given frequency rule. Then the implied GFT is interpolated for fixed maturities.
}
