% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceImpliedGFT.R
\name{surfaceImpliedGFT}
\alias{surfaceImpliedGFT}
\title{Calculate the implied Laplace transform using a joint fitting of a smooth surface to all options.}
\usage{
surfaceImpliedGFT(option.panels, mkt.frame, u.t.mat, max.pow = 3,
  doPlot = FALSE, doFitPlot = 0, verbose = FALSE, spline.rank = 6,
  convergence.scale = 0.01, weights = "opt-sqrt")
}
\arguments{
\item{option.panels}{A list, each element containing a dataframe with normalized mid OTM option prices and LOG strikes. Each panel should correspond to the SAME day.}

\item{u.t.mat}{An Nx2 dataframe, consisting of the frequencies and maturities for which the laplace transforms should be calculated}

\item{doPlot}{Whether fitted / true values should be plotted}

\item{doFitPlot}{Whether backfitting iterations should provide plot, \code{doFitPlot} is an integer: if 0, not plots, otherwise plots every \code{doFitPlot} iterations}

\item{verbose}{Do we want diagnostic information?}

\item{spline.rank}{This is the \code{k} argument to the function \code{\link{s}} from package \link{mgcv}, it determines the spline basis dimension.}

\item{convergence.scale}{Scaling of the convergence criterion for the backfitting algorithm, a submodel \code{j} is deemed as having converged if the sqrt-sample-size-scaled norm of the iteration change in the predicted values is smaller than \code{convergence.scale * sd(cp - sum(k != j)fk)}}

\item{weights}{string which defines the weighting scheme to be used in fitting. \code{opt-sqrt} weighs by the square root of the out of the money option price.}

\item{mkt.list}{A data-frame, where each row corresponds to the market features associated with the n-th option panel (e.g. maturity, interest rate, etc...)}
}
\value{
An Nx3 dataframe, where the last column is the calculated transform value
}
\description{
Calculate the implied Laplace transform using a joint fitting of a smooth surface to all options.
}
