% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpsImpliedDivergence.R
\name{tpsImpliedDivergence}
\alias{tpsImpliedDivergence}
\title{Calculate the implied divergences from option panels}
\usage{
tpsImpliedDivergence(option.panels, mkt.frame, u.t.mat, doPlot = FALSE,
  doFitPlot = 0, verbose = FALSE, L = NULL, U = NULL, bootstrap = F,
  Nrepl = 1000, ...)
}
\arguments{
\item{option.panels}{A list, each element containing a dataframe with normalized mid OTM option prices and LOG strikes. Each panel should correspond to the SAME day.}

\item{u.t.mat}{An Nx3 dataframe, consisting of the powers, maturities, and divergence types}

\item{verbose}{Do we want diagnostic information?}

\item{...}{further arguments to be passed to \code{\link{panelSmoother}} and further to \code{\link{rgl::persp3d}}}

\item{mkt.list}{A data-frame, where each row corresponds to the market features associated with the n-th option panel (e.g. maturity, interest rate, etc...)}

\item{spline.rank}{This is the \code{k} argument to the function \code{\link{s}} from package \link{mgcv}, it determines the spline basis dimension.}

\item{convergence.scale}{Scaling of the convergence criterion for the backfitting algorithm, a submodel \code{j} is deemed as having converged if the sqrt-sample-size-scaled norm of the iteration change in the predicted values is smaller than \code{convergence.scale * sd(cp - sum(k != j)fk)}}
}
\value{
An Nx3 dataframe, where the last column is the calculated transform value
}
\description{
Calculate the implied divergences from option panels
}
